+++ reco/src/main/java/com/example/reco/services/RatingService.java
@@ -2,6 +2,7 @@ package com.example.reco.services;
 
 import com.example.reco.controllers.dto.RatingResponse;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface RatingService {
@@ -9,12 +10,12 @@ public interface RatingService {
 
     RatingResponse getRatingById(Long id);
 
-    RatingResponse rateItem(Long item_id, Long user_id, Short grade);
+    RatingResponse rateItem(Long item_id, Long user_id, BigDecimal grade);
 
     List<RatingResponse> getRatingsByUser(Long userId, int limit);
 
     List<RatingResponse> getRatingsByItem(Long itemId, int limit);
 
-    RatingResponse updateRating(Long id, Short newGrade);
+    RatingResponse updateRating(Long id, BigDecimal newGrade);
 
 }
